% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{add_groups}
\alias{add_groups}
\title{Add higher level groups to a taxonomy table.}
\usage{
add_groups(x)
}
\arguments{
\item{x}{A taxonomy table.}
}
\value{
A taxonomy table with higher level groups added.
}
\description{
Add higher level groups to a taxonomy table.
}
