% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_databases.R
\name{generate_reference_species}
\alias{generate_reference_species}
\title{Create species lists with WoRMS accepted names for a set of reference databases.
Results are written to a compressed CSV file in the workspace.}
\usage{
generate_reference_species(
  reference_databases,
  workspace = system.file("data", package = "ednagaps")
)
}
\arguments{
\item{reference_databases}{A named list of reference databases, with at least taxon table paths.}

\item{workspace}{Path to the workspace.}
}
\value{
Output path.
}
\description{
Create species lists with WoRMS accepted names for a set of reference databases.
Results are written to a compressed CSV file in the workspace.
}
