% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_databases.R
\name{get_marker_species_list}
\alias{get_marker_species_list}
\title{Get a list of species names as provided from a reference database.}
\usage{
get_marker_species_list(reference_database_path, marker)
}
\arguments{
\item{marker}{Marker name.}

\item{reference_database_taxa_path}{Reference database taxon table path.}
}
\description{
Get a list of species names as provided from a reference database.
}
