% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\docType{data}
\name{groups}
\alias{groups}
\title{Higher level groups for analysis purposes.}
\format{
An object of class \code{list} of length 19.
}
\usage{
groups
}
\description{
Higher level groups for analysis purposes.
}
\keyword{datasets}
