% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_databases.R
\name{read_reference_species}
\alias{read_reference_species}
\title{Read species lists with WoRMS accepted names from the compressed CSV file in the workspace.}
\usage{
read_reference_species(workspace = system.file("data", package = "ednagaps"))
}
\arguments{
\item{workspace}{Path to the workspace.}
}
\description{
Read species lists with WoRMS accepted names from the compressed CSV file in the workspace.
}
