% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_databases.R
\name{worms_for_names}
\alias{worms_for_names}
\title{Match names with WoRMS.}
\usage{
worms_for_names(...)
}
\arguments{
\item{x}{A character vector of names.}
}
\value{
A data frame with the input names and their WoRMS matches.
}
\description{
Match names with WoRMS.
}
